+++ src/MarginTrading.Common/BackendContracts/GraphBidAskPairBackendContract.cs
@@ -4,8 +4,8 @@ namespace MarginTrading.Common.BackendContracts
 {
     public class GraphBidAskPairBackendContract
     {
-        public double Bid { get; set; }
-        public double Ask { get; set; }
+        public decimal Bid { get; set; }
+        public decimal Ask { get; set; }
         public DateTime Date { get; set; }
     }
 }
