+++ src/MarginTrading.Core/IQuoteHistoryRepository.cs
@@ -4,6 +4,6 @@ namespace MarginTrading.Core
 {
 	public interface IQuoteHistoryRepository
 	{
-		Task<double?> GetClosestQuoteAsync(string instrument, OrderDirection direction, long ticks);
+		Task<decimal?> GetClosestQuoteAsync(string instrument, OrderDirection direction, long ticks);
 	}
 }
